Editor: Flatten BlockListLayout into base BlockList
